% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file_or_key}
\alias{is_file_or_key}
\title{Determine if a string is a file path or a ChatGPT API key}
\usage{
is_file_or_key(x)
}
\arguments{
\item{x}{A character string to be tested.}
}
\value{
A character string indicating the type: 
  \itemize{
    \item `"file"` if the input is an existing file path.
    \item `"chatgpt_key"` if the input matches the OpenAI API key pattern.
    \item `"unknown"` otherwise.
  }
}
\description{
This function checks whether the provided string corresponds to an 
existing file path, matches the pattern of an OpenAI ChatGPT API key, 
or is unknown.
}
\examples{
is_file_or_key("path/to/myfile.txt")
is_file_or_key("sk-abcdefghijklmnopqrstuvwxyz1234567890abcd")
is_file_or_key("randomstring")

}
